# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: iROC GmbH website API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned GCS upload URL (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
...
  /training/register:
    post:
      operationId: registerForTraining
      tags: [training]
      summary: Register a doctor for a training session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingRegistrationInput"
      responses:
        "201":
          description: Training registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/training-registrations/{id}/confirm:
    post:
      operationId: confirmTrainingRegistration
      tags: [admin]
      summary: Admin - confirm a training registration and synchronize its lead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration confirmed and lead synchronized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmTrainingRegistrationResponse"
        "400":
          description: Invalid registration ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Registration cannot be confirmed from its current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Registration is missing required synchronization data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TrainingRegistrationInput:
      type: object
      required:
        - salutation
        - medicalDegree
        - firstName
        - lastName
        - specialty
        - institutionName
        - address
        - postalCode
        - city
        - country
        - phone
        - email
        - trainingDateId
        - instrument
        - privacyConsent
        - marketingConsent
      properties:
        salutation:
          type: string
          enum: [Mann, Frau, Diverse, Andere]
        medicalDegree:
          type: string
          enum: [Dr., "Dr. med.", Other]
        firstName:
          type: string
        lastName:
          type: string
        specialty:
          type: string
        institutionName:
          type: string
        address:
          type: string
        street:
          type: ["string", "null"]
        houseNumber:
          type: ["string", "null"]
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        fax:
          type: ["string", "null"]
        email:
          type: string
        trainingDateId:
          type: integer
        instrument:
          type: string
          enum: [spirecut, ministem]
        websiteUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        privacyConsent:
          type: boolean
        marketingConsent:
          type: boolean
...
    RequestUploadUrlBody:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    RequestUploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
...
    ConfirmTrainingRegistrationResponse:
      type: object
      required: [registrationId, status, confirmedAt, leadId, leadCreated]
      properties:
        registrationId:
          type: integer
        status:
          type: string
          enum: [confirmed]
        confirmedAt:
          type: ["string", "null"]
        leadId:
          type: integer
        leadCreated:
          type: boolean
...